move keywords to script
