+++ chatsky/core/__init__.py
@@ -6,4 +6,4 @@ from chatsky.core.script import Node, Flow, Script
 from chatsky.core.script_function import BaseCondition, BaseResponse, BaseDestination, BaseProcessing, BasePriority
 from chatsky.core.transition import Transition
 from chatsky.core.node_label import NodeLabel, AbsoluteNodeLabel
-from chatsky.core.keywords import GLOBAL, LOCAL, RESPONSE, TRANSITIONS, MISC, PRE_RESPONSE, PRE_TRANSITION
+from chatsky.core.script import GLOBAL, LOCAL, RESPONSE, TRANSITIONS, MISC, PRE_RESPONSE, PRE_TRANSITION
